Organise Workflow stores Move `Workflow` into the `WorkflowStore` directory.
